added registration and login features with firebase
