+++ package.json
@@ -9,12 +9,17 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.2",
+    "@react-navigation/native-stack": "^6.9.8",
     "expo": "^47.0.12",
     "expo-status-bar": "~1.4.2",
+    "firebase": "^9.15.0",
     "react": "^18.1.0",
     "react-devtools": "^4.27.1",
     "react-native": "0.70.5",
     "react-native-reanimated": "~2.12.0",
+    "react-native-safe-area-context": "4.4.1",
+    "react-native-screens": "~3.18.0",
     "react-native-svg": "13.4.0"
   },
   "devDependencies": {
