+++ screens/HomeScreen.js
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const HomeScreen = () => {
+  return (
+    <View>
+      <Text>Home Screen</Text>
+    </View>
+  )
+}
+
+export default HomeScreen
+
+const styles = StyleSheet.create({})
